fix(cli): fix stack monitoring when the stack events do not have phsical resource id set (#27692)
`PhysicalResourceId` for a `StackEvent` is [not required](https://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_StackEvent.html) and can be empty. This causes error in stack deployment activty monitoring where we try to monitor nested stacks but empty `PhysicalResourceId` in the `StackEvent` causes it to fail with the error

```
[ValidationError]: 2 validation errors detected: 
Value '' at 'stackName' failed to satisfy constraint: Member must have length greater than or equal to 1; 
Value '' at 'stackName' failed to satisfy constraint: Member must satisfy regular expression pattern: [a-zA-Z][-a-zA-Z0-9]*|arn:[-a-zA-Z0-9:/._+]*
```
This PR adds a check to avoid that.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*