feat(cloudformation-diff): use awscdk-service-spec as data source (#27813)
We already changed the datasource for generated resource to use `@aws-cdk/awscdk-service-spec`.
However `cloudformation-diff` was still using the old data source. 

This PR swaps out the source, with minimal further changes.
Hence no additional tests are required.

Note that this new implementation does not apply `temporary-schemas` defined in `spec2cdk`.
This is because `temporary-schemas` is a mechanism to unblock development of unreleased features, and not a mechanism to publish schema changes.
For the latter, changes must be made in [cdklabs/awscdk-service-spec](https://github.com/cdklabs/awscdk-service-spec).

I ran a local benchmark and `cdk diff` is now running ~1.03 slower than before.
I investigated why and this appears to be mostly due to the gzip compression of the database file.
If we load the DB from an uncompressed source, the new code path performs ~1.07 times faster than the old code path. This trade-off is acceptable.

Compressed package size of `aws-cdk` is reducing with this change from `11.7MB` to `8.9MB`

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*