chore(prlint): don't mark PRs as 'not ready' on community comment (#27819)
Community reviewers have the ability to choose any of the
approve/comment/request changes buttons that are available in the review
tab. Prior to this change, `prlint` would consider a comment from a
community reviewer as being equivalent to requesting changes (and in
fact, didn't consider the requesting changes case). This would remove
the `pr/needs-community-review` label which surprised some reviewers.

With this change, `prlint` will only remove the
`pr/needs-community-review` label when a community reviewer specifically
chooses "request changes". Additionally, reviewers are now able to
switch from approving to requesting changes (this doesn't override any
other reviewers' approvals, just that reviewer's own).

Additionally, this adds mocks for the `getTrustedCommunityMembers`
method and avoids hardcoding the logins of multiple community reviewers
into the tests. As a side effect, the tests also run faster since `curl`
isn't being invoked so frequently.

Here is a screenshot of what I see when reviewing on this repo:
![image](https://github.com/aws/aws-cdk/assets/850893/03d3b5e5-2798-47bf-951d-b72964f974aa)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*