chore(spec2cdk): fix cli not working for single services (#27892)
Fixes an issue with the CLI version of `spec2cdk` not working for single services anymore.
This used to be implicitly tested by being included in the hot path. However after a refactor, we were not using it anymore and thus didn't notice the breaking. Added a test to ensure future functionality.

Replaced to custom options parser with Node's built-in parser for increased usability and support for list options (calling the same option multiple time).

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*