fix(eks): add neuron plugin in addNodegroupCapacity()
Previously, the neuron plugin was only added in addAutoScalingGroupCapacity() when
the instanceType was some of the INFERENTIA types.

Let's also add the plugin if at least one (some/any) of the
instanceTypes in addNodegroupCapacity() is of the INFERENTIA types.

added integration tests for the addNodegroupCapacity() feature too,
and snapshot test

Co-authored-by: Luca Pizzini <lpizzini7@gmail.com>