fix(elasticloadbalancingv2): access logging fails on imported bucket (#27948)
https://github.com/aws/aws-cdk/pull/27558 fixed a circular dependency issue between the bucket and the ELB when using `logAccessLogs` by adding the dependency at the L1 level.
The dependency was added on both the bucket and the bucket policy (which already refers to the bucket).

This introduced a new bug when adding a dependency for imported buckets (as `bucket.node.defaultChild` is `undefined`).

This PR fixes the issue by adding the dependency at the L1 level on the bucket policy directly.

Closes #27928.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*