chore: pr-build failing with unsafe-perm is not a valid npm option (#28068)
We switched the images used to build PRs to `jsii/superchain:bullseye-slim-node18` which comes with Node.js 18 instead of Node.js 16. This Node.js upgrade also includes an upgraded version of npm, and in recent npm versions, the unsef-perm option is not required anymore. This change simply removes the now redundant command.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*