chore(core): use key instead of logicalId for key when provided (#28075)
In CDK Migrate, when a user has a construct with the same name as their CfnOutput key, it creates a naming collision.

This change enables us to name the construct something else and still use the correct key in creating the output. Note that the logicalId only makes it into the synthesized template as the output key so this change won't impact existing apps using the CfnOutput construct.

Closes #<issue number here>.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*