fix(s3-notifications): multiple notifications doesn't work
When multiple bucket notifications are created it creates a race
condition where only the last one processed gets applied. All bucket
notifications created in a stack are given the same `stackId` prefix.
This prefix is then used to filter out the notification created by the
custom resource. If there are other notifications created in the same
stack, but not by this custom resource, they get filtered out.

This PR fixes that by filtering the notifications by the specific
notification id. This ensures that only the notifications created by the
individual custom resource are filter out and the rest (included those
created by other custom resources) are marked external.