fix(spec2cdk): attributes in `ICfnResource` should not be optional (#28179)
They should all be required. Anything returned from `ref` will exist,
even if the property itself is not required. For example,
`CfnBucket.attrBucketName` is not required in the resource spec because
you can either specify a bucket name or have one generated for you.
However, when accessing it via `ref`, the bucket name will certainly
exist, so it is not optional.

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*