fix(apigatewayv2): export newly graduated apigatewayv2 modules (#28250)
The API Gateway V2 modules which were graduated to stable with [2.112.0](https://github.com/aws/aws-cdk/releases/tag/v2.112.0) were not added to the main package exports. In particular:

- aws-apigatewayv2-authorizers
- aws-apigatewayv2-integrations

Closes #28239.

Now, I am not completely familiar with JSII and the CDK, so this may be only one part of the required solution to include these missing packages. 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*