fix(apigateway): move url property to RestApiBase (#27742)
Moved the url property from RestApi class to the RestApiBase to have the url property on all derived class. This will fix the problem mentioned in the issue where SpecRestApi does not work with the RestApiOrigin construct.

PS. I could not perform the integration tests as the free tier on my personal account expired.

Closes #27501.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*