Add support for imported alias grants
This will add a new properties to the AliasAttributes, making it possible to import by aliasArn or cross account alias by aliasName with account and region.

A new static method called fromAliasArn is a shortcut to fromAliasAttribute with aliasArn specified.

Granting access to the alias will create a kms:ResourceAliases condition to the principal if no target key is available.

The IAlias interface was missing the aliasArn property