feat(ecs): Support specifying revision of task definition (#27036)
If using CodePipeline EcsDeployAction without using the CODE_DEPLOY deployment controller, future deployments of an ECS service will revert the task definition to the task definition deployed by CloudFormation, even though the latest active revision created by the deploy action is the one that is intended to be used. This provides a way to specify the specific revision of a task definition that should be used, including the special value `latest` which uses the latest ACTIVE revision.

Closes #26983.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*