feat(ecs-patterns): support disabling CPU-based scaling and custom target utilization (#28315)
Added an optional parameter that defaults to false over the CPU-based scaling policy that conflicts with the queue visible message-based policy.

When disabled this will stop the race condition issue mentioned in #20706 by only allowing the scaling of the number of messages on the queue similar to the SQS-Lambda pattern.

Note: If this parameter is enabled then this bug will crop up again and the user has to handle the container termination manually.

Updated integration tests and unit tests are working.

Closes #20706 .

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*