doc(core): update old recommendations in core README
`splitArn` is the correct updated version of `parseArn`. It's now
necessary to use `ArnFormat` not a literal string for the format of the
ARN resource for `splitArn` and `formatArn`.

Finally, to fit with the recommendations to use a singleton pattern for
custom resource handlers, the docs for Lambda are changed to recommend
just using a `SingletonFunction`.