feat(stepfunctions-tasks): add `timeout` parameter for EmrCreateCluster (#28532)
This PR adds a new parameter `timeout` as Duration type instead of `timeoutDurationMinutes` because the `timeoutDurationMinutes` is a number type.

Originally, `timeoutDurationMinutes` was a **required** parameter, but we have made it **optional** and also made the new parameter **optional** to avoid breaking change.

Instead, added a validation to ensure that the value is specified.

We discussed this in the following thread: https://github.com/aws/aws-cdk/pull/28529#discussion_r1438587964

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*