feat(ecs): `interactive` option in `ContainerDefinitionOptions` (#28536)
This pull request adds the `interactive` argument to `ContainerDefinitionOptions`. This argument is used when deploying containerized applications that require the allocation of standard input (stdin) or a terminal (tty). This parameter corresponds to `OpenStdin` in the "Create a container" section of the Docker Remote API and the `--interactive` option to `docker run`.

Closes #24326.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*