feat(rds): `timeout` and `timeoutAction` properties to ServerlessCluster (#28534)
This pull request introduces two new properties to the `ServerlessCluster` class in the AWS CDK RDS package: `secondsBeforeTimeout` and `timeoutAction`. 

The `secondsBeforeTimeout` property allows users to specify the amount of time that Aurora Serverless v1 will attempt to find a scaling point to perform seamless scaling before enforcing the timeout action. The default value is 300 seconds (5 minutes).

The `timeoutAction` property allows users to specify the action to take when the timeout is reached. Users can choose between `ForceApplyCapacityChange`, which will force the capacity to the specified value as soon as possible, even without a scaling point, and `RollbackCapacityChange`, which will ignore the capacity change if a scaling point is not found. The default behavior is `RollbackCapacityChange`.

These enhancements provide users with more control over the scaling behavior of their Aurora Serverless clusters.

Closes #27183

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*