fix(stepfunctions): retry block in CustomState is always empty (#28598)
Currently it is not possible to set Retry for the Task when using CustomState construct.
There are two ways to add Retry: by adding the `addRetry` method or by rendering the `Retry` property of `stateJson`.
The `addRetry` method was added in this PR because rendering the `Retry` property of `stateJson`, which was not rendered before, could result in an unexpected StateMachine for the user.

Closes #28586

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*