fix(cloudfront): function name is non-deterministic
This fixes an issue where having a long stack name or resource ID could
result in inconsistent naming of the function (replacing it on each
deployment) because the length of the token that holds the region may
not be consistent. This uses a more deterministic method based on
checking whether the region is resolved and using the
`uniqueResourceName` function.

It is important to note that while this doesn't change the name of any
current resources in the unit/integration tests, I am not 100% certain
that this won't impact any existing users.