chore: auto-sync every fork from upstream (#28653)
Add a GitHub action that will update the current repository from upstream on a daily basis.

This makes it so that various forks of this repository automatically keep themselves up-to-date with the parent repo, and it will be that much easier to make PRs off of a recent, up-to-date clone, without having to do additional manual syncing.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*